requirement.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.1
mysqlclient==2.2.0
requests==2.31.0